+++ fp-bindgen/src/types/structs.rs
@@ -2,7 +2,7 @@ use super::TypeIdent;
 use crate::types::format_bounds;
 use crate::{casing::Casing, docs::get_doc_lines};
 use quote::ToTokens;
-use std::{collections::BTreeMap, convert::TryFrom};
+use std::convert::TryFrom;
 use syn::{
     ext::IdentExt, parenthesized, parse::Parse, parse::ParseStream, Attribute, Error, GenericParam,
     Ident, ItemStruct, LitStr, Result, Token,
@@ -56,17 +56,17 @@ pub(crate) fn parse_struct_item(item: ItemStruct) -> Struct {
 pub struct StructOptions {
     pub field_casing: Casing,
 
-    /// Rust module paths where the type can be found for the given generator.
+    /// Rust module path where the type can be found for the given generator.
     /// If present, the generator can use this type instead of generating it.
     ///
     /// ## Example:
     ///
     /// ```rs
-    /// #[fp(rust_plugin_module = "my_crate")]
+    /// #[fp(rust_module = "my_crate")]
     /// struct MyStruct { /* ... */ }
     /// ```
     ///
-    /// This will insert `"rust_plugin" => "my_crate"` into the map, which can
+    /// This will set `"my_crate"` as the rust_module, to
     /// be used by the Rust plugin generator to generate a `use` statement such
     /// as:
     ///
@@ -75,7 +75,7 @@ pub struct StructOptions {
     /// ```
     ///
     /// Instead of generating the struct definition itself.
-    pub native_modules: BTreeMap<String, String>,
+    pub rust_module: Option<String>,
 }
 
 impl StructOptions {
@@ -95,8 +95,8 @@ impl StructOptions {
         if other.field_casing != Casing::default() {
             self.field_casing = other.field_casing;
         }
-        for (key, value) in other.native_modules.iter() {
-            self.native_modules.insert(key.clone(), value.clone());
+        if let Some(other_rust_module) = &other.rust_module {
+            self.rust_module = Some(other_rust_module.clone());
         }
     }
 
@@ -132,10 +132,8 @@ impl Parse for StructOptions {
                     result.field_casing = Casing::try_from(parse_value()?.as_ref())
                         .map_err(|err| Error::new(content.span(), err))?
                 }
-                module if module.ends_with("_module") => {
-                    result
-                        .native_modules
-                        .insert(module[0..module.len() - 7].to_owned(), parse_value()?);
+                "rust_module" => {
+                    result.rust_module = Some(parse_value()?);
                 }
                 other => {
                     return Err(Error::new(